Add the pagination for the main and category pages
Changes:
Add the pagination{
Add a php document of my custom pagination.
Update the layout for show the pagination.
Call the pagination with the "links()" sentence in the main and category
views.
Add the route for show the page with category filter.
}
Add a function in the "Entrada" model{
Add the "scopeDateDescending" function to get the "posts" descendent
sorted by date.
}